refactor: sockets
