+++ src/clients/createClient.test.ts
@@ -141,6 +141,7 @@ describe('transports', () => {
         "request": [Function],
         "transport": {
           "getSocket": [Function],
+          "getSocketClient": [Function],
           "key": "webSocket",
           "name": "WebSocket JSON-RPC",
           "request": [Function],
