+++ src/clients/createPublicClient.test.ts
@@ -308,6 +308,7 @@ describe('transports', () => {
         "simulateContract": [Function],
         "transport": {
           "getSocket": [Function],
+          "getSocketClient": [Function],
           "key": "webSocket",
           "name": "WebSocket JSON-RPC",
           "request": [Function],
