+++ src/clients/createTestClient.test.ts
@@ -256,6 +256,7 @@ describe('transports', () => {
         "stopImpersonatingAccount": [Function],
         "transport": {
           "getSocket": [Function],
+          "getSocketClient": [Function],
           "key": "webSocket",
           "name": "WebSocket JSON-RPC",
           "request": [Function],
