+++ src/utils/index.test.ts
@@ -26,6 +26,9 @@ test('exports utils', () => {
       "concatBytes": [Function],
       "concatHex": [Function],
       "containsNodeError": [Function],
+      "createHttpClient": [Function],
+      "createSocketClient": [Function],
+      "createWebSocketClient": [Function],
       "decodeAbiParameters": [Function],
       "decodeErrorResult": [Function],
       "decodeEventLog": [Function],
@@ -152,6 +155,7 @@ test('exports utils', () => {
       "slice": [Function],
       "sliceBytes": [Function],
       "sliceHex": [Function],
+      "socketClientCache": Map {},
       "stringToBytes": [Function],
       "stringToHex": [Function],
       "stringify": [Function],
