+++ src/utils/rpc/id.ts
@@ -1,0 +1,13 @@
+export function createId() {
+  return {
+    current: 0,
+    take() {
+      return this.current++
+    },
+    reset() {
+      this.current = 0
+    },
+  }
+}
+
+export const idCache = createId()
