+++ test/setup.ts
@@ -5,7 +5,7 @@ import { afterAll, afterEach, beforeAll, beforeEach, vi } from 'vitest'
 import { setIntervalMining } from '~viem/actions/test/setIntervalMining.js'
 import { cleanupCache, listenersCache } from '~viem/utils/observe.js'
 import { promiseCache, responseCache } from '~viem/utils/promise/withCache.js'
-import { socketsCache } from '~viem/utils/rpc.js'
+import { socketClientCache } from '~viem/utils/rpc/socket.js'
 
 import { forkBlockNumber, poolId } from './src/constants.js'
 import { setBlockNumber, testClient } from './src/utils.js'
@@ -25,7 +25,7 @@ beforeEach(async () => {
   responseCache.clear()
   listenersCache.clear()
   cleanupCache.clear()
-  socketsCache.clear()
+  socketClientCache.clear()
 
   if (process.env.SKIP_GLOBAL_SETUP) return
   await setIntervalMining(testClient, { interval: 0 })
