+++ test/src/utils.ts
@@ -22,7 +22,7 @@ import { RpcRequestError } from '~viem/errors/request.js'
 import type { Chain } from '~viem/types/chain.js'
 import { type EIP1193Provider, ProviderRpcError } from '~viem/types/eip1193.js'
 import { namehash } from '~viem/utils/ens/namehash.js'
-import { rpc } from '~viem/utils/rpc.js'
+import { createHttpClient } from '~viem/utils/rpc/http.js'
 
 import { type RequestListener, createServer } from 'http'
 import type { AddressInfo } from 'net'
@@ -105,7 +105,8 @@ const provider: EIP1193Provider = {
         },
       ]
 
-    const { error, result } = await rpc.http(localHttpUrl, {
+    const httpClient = createHttpClient(localHttpUrl)
+    const { error, result } = await httpClient.request({
       body: {
         method,
         params,
